requirements.txt
pandas
spacy
textblob
scikit-learn
sentence_transformers
faiss-cpu
streamlit
plotly
dask[complete]